Add invites, challenges, savings-growlio sync, and notification read tracking
Merge budget into cashflow plan items, add spousal invite flow, savings
challenges, growlio account sync for savings products, per-user
notification read state, and coaching threshold settings. Consolidate
run.sh/run.bat into dev.sh/dev.bat and add PWA manifest/icons.

+++ app/models/account.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from sqlalchemy import Boolean, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -10,6 +12,6 @@ class Account(Base):
     id: Mapped[int] = mapped_column(primary_key=True)
     name: Mapped[str] = mapped_column(String(100))
     account_type: Mapped[str] = mapped_column(String(10))  # 'bank' | 'cash' | 'card'
-    initial_balance: Mapped[float] = mapped_column(Numeric(14, 2), default=0)
+    initial_balance: Mapped[Decimal] = mapped_column(Numeric(14, 2), default=0)
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
     sort_order: Mapped[int] = mapped_column(Integer, default=0)
